feat: Add CategorySelector, FeedbackFormSection, and Project Management components
- Implemented CategorySelector for selecting multiple categories with animations.
- Created FeedbackFormSection to handle feedback submission with project and recipient selection.
- Developed ProjectCard for displaying project details and managing user assignments.
- Introduced ProjectManagementSection to manage multiple projects and their assignments.
- Added RatingSelector for user rating input with visual feedback.
- Implemented RecipientSelector for selecting feedback recipients based on project.
- Created UserAssignmentCard for managing user assignments within projects.
- Developed UserDropdown for selecting users with visual indicators of workload.
